# Copy less arrays. Improved performance to 12.7 seconds (-2 seconds)