```
hyperfine --warmup 1 --runs 5 --export-json me-timing.json ./calculate_average_bytesfellow.sh
Benchmark 1: ./calculate_average_bytesfellow.sh
  Time (mean ± σ):     12.756 s ±  0.501 s    [User: 82.904 s, System: 5.002 s]
  Range (min … max):   11.952 s … 13.299 s    5 runs
```